apps/backend/requirements.txt
fastapi>=0.135.0
uvicorn>=0.44.0
sqlalchemy>=2.0.49
asyncpg>=0.31.0
pydantic-settings>=2.13.0
alembic>=1.18.0
python-jose>=3.5.0
passlib>=1.7.4
bcrypt>=5.0.0
python-multipart>=0.0.20
authlib>=1.5.0
httpx>=0.28.0
requests>=2.32.0
PyJWT>=2.8.0
cryptography>=42.0.0
boto3>=1.34.0
pytest>=9.0.0
pytest-asyncio>=1.3.0
factory-boy>=3.3.0
